chore: 공통 규칙에 따른 변경
